Edit user for admin menu refactored
